started pilot tutorial documentation for R users
Downloaded the R markdown template for the Geospatial Workbook and started adding details for a Pilot Tutorial. It's also a kind of testing ground for planning the R package, so some code lines will likely move unless we want to show how to play with the web api from R but not with our R package.